+++ quam_builder/architecture/superconducting/qpu/flux_tunable_quam.py
@@ -130,6 +130,9 @@ class FluxTunableQuam(BaseQuam):
         """Initialize the QPU with the calibrated TWPA pumping points and
            with the specified flux point and target
 
+        The TWPA pumps require calling :meth:`twpa_keepalive` once per
+        inner-most-loop iteration to keep them on.
+
         Args:
             flux_point (str): The flux point to set. Default is 'joint'.
             target: The qubit under study.